backend/requirements.txt
Flask==2.2.3
Flask-Login==0.6.2
pymongo==4.3.3
pymongo
flask_cors
textblob
nltk
werkzeug==2.2.2
pandas
scikit-learn
bs4
psycopg2
python-dotenv
matplotlib
Unidecode
geonamescache
dateutils
casefy
lxml
regex
botasaurus
Flask_session
spacy
SQLAlchemy
# python -m spacy download en_core_web_sm










